# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Slot Game Backend API
  version: 0.1.0
  description: |
    老虎机游戏后端 API 文档（基于当前代码实现整理）。
    注意：部分响应结构含有嵌套对象/动态字段，此处以通用对象占位。
servers:
  - url: http://localhost:8080
    description: Local Dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BalanceResponse:
      type: object
      properties:
        balance:
          type: integer
        frozen_amount:
          type: integer
        available:
          type: integer
        currency:
          type: string
          example: COIN
    DepositRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          minimum: 100
    DepositResponse:
      type: object
      properties:
        transaction_id:
          type: string
        amount:
          type: integer
        balance:
          type: integer
        status:
          type: string
        created_at:
          type: string
          format: date-time
    WithdrawRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          minimum: 100
    WithdrawResponse:
      allOf:
        - $ref: '#/components/schemas/DepositResponse'
    TransactionInfo:
      type: object
      properties:
        id:
          type: integer
        order_no:
          type: string
        type:
          type: string
          description: bet/win/deposit/withdraw
        amount:
          type: integer
        before_balance:
          type: integer
        after_balance:
          type: integer
        description:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionInfo'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
paths:
...
  /api/v1/wallet/balance:
    get:
      tags: [Wallet]
      summary: 获取余额
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /api/v1/wallet/deposit:
    post:
      tags: [Wallet]
      summary: 充值（测试）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
  /api/v1/wallet/withdraw:
    post:
      tags: [Wallet]
      summary: 提现（模拟）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
  /api/v1/wallet/transactions:
    get:
      tags: [Wallet]
      summary: 交易记录
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
            description: bet/win/deposit/withdraw
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
  /api/v1/wallet/statistics:
    get:
      tags: [Wallet]
      summary: 钱包统计
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          schema:
            type: string
            example: 2025-09-10
      responses:
        '200':
          description: OK
...